Move header to /purchases
